+++ src/utils/typed_array.cpp
@@ -88,7 +88,7 @@ Local<Value> TypedArray::New(GDALDataType type, void *data, int64_t length) {
     case GDT_Int32: name = "Int32Array"; break;
     case GDT_UInt32: name = "Uint32Array"; break;
 #if GDAL_VERSION_MAJOR > 3 || (GDAL_VERSION_MAJOR == 3 && GDAL_VERSION_MINOR >= 11)
-    case GDT_Float16: name = "FLoat16Array"; break;
+    case GDT_Float16: throw "Float16Array cannot be created at the moment, pass an existing array"; break;
 #endif
     case GDT_Float32: name = "Float32Array"; break;
     case GDT_Float64: name = "Float64Array"; break;